+++ src/Vankosoft/UsersBundle/DependencyInjection/Configuration.php
@@ -7,12 +7,18 @@ use Symfony\Component\Config\Definition\Builder\ArrayNodeDefinition;
 use Sylius\Bundle\ResourceBundle\SyliusResourceBundle;
 use Sylius\Component\Resource\Model\ResourceInterface;
 use Sylius\Component\Resource\Factory\Factory;
+use Sylius\Bundle\ResourceBundle\Doctrine\ORM\EntityRepository;
 
 use VS\UsersBundle\Model\User;
 use VS\UsersBundle\Model\UserInterface;
 use VS\UsersBundle\Controller\UsersController;
 use VS\UsersBundle\Repository\UsersRepository;
 use VS\UsersBundle\Form\UserFormType;
+
+use VS\UsersBundle\Model\UserRole;
+use VS\UsersBundle\Model\UserRoleInterface;
+use VS\UsersBundle\Controller\UsersRolesController;
+use VS\UsersBundle\Form\UserRoleForm;
 
 use VS\UsersBundle\Model\UserInfo;
 use VS\UsersBundle\Model\UserInfoInterface;
@@ -77,6 +83,27 @@ class Configuration implements ConfigurationInterface
                             ->end()
                         ->end()
                         // End Users
+                        
+                        // Begin User Roles
+                        ->arrayNode( 'user_roles' )
+                            ->addDefaultsIfNotSet()
+                            ->children()
+                                ->variableNode( 'options' )->end()
+                                ->arrayNode( 'classes' )
+                                    ->addDefaultsIfNotSet()
+                                    ->children()
+                                        ->scalarNode( 'model' )->defaultValue( UserRole::class )->cannotBeEmpty()->end()
+                                        ->scalarNode( 'interface' )->defaultValue( UserRoleInterface::class )->cannotBeEmpty()->end()
+                                        ->scalarNode( 'controller' )->defaultValue( UsersRolesController::class )->cannotBeEmpty()->end()
+                                        //->scalarNode( 'repository' )->cannotBeEmpty()->end()
+                                        ->scalarNode( 'repository' )->defaultValue( EntityRepository::class )->cannotBeEmpty()->end()
+                                        ->scalarNode( 'factory' )->defaultValue( Factory::class )->cannotBeEmpty()->end()
+                                        ->scalarNode( 'form' )->defaultValue( UserRoleForm::class )->cannotBeEmpty()->end()
+                                    ->end()
+                                ->end()
+                            ->end()
+                        ->end()
+                        // End User Roles
                         
                         // Begin UserInfo
                         ->arrayNode('user_info')